+++ src/ResourceManager/Network/Commands.Network/ChangeLog.md
@@ -44,6 +44,7 @@
 
 ## Version 6.7.0
 * Updated cmdlet New-AzureRmVirtualNetworkGatewayConnection with support for switch ExpressRouteGatewayBypass
+* Added cmdlets for Azure SdWan
 
 ## Version 6.6.1
 * Changed default cmdlet output presentation to table view